requirements.txt
adb-shell
pywebview
Flask
dotmap